[50] Refactor PlainTextHelperView code
