+++ Sources/CBLNotes/Features/Interpreter/Commands/Markdown/PlainText/Models/TextPart.swift
