+++ Sources/CBLNotes/Features/Interpreter/Commands/Markdown/PlainText/Models/TextStyle.swift
